pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docswarm"
version = "0.1.0"
description = "Transform PDF corpora into structured wikis using an AI agent swarm"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "duckdb>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["docswarm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
